crud produto e categoria atualizado

+++ src/main/java/com/generation/bigfarma/controller/ProdutoController.java
@@ -1,8 +1,12 @@
 package com.generation.bigfarma.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,36 +14,66 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.generation.bigfarma.model.Produto;
 import com.generation.bigfarma.repository.ProdutoRepository;
 
 
+import jakarta.validation.Valid;
+
 @RestController
 @RequestMapping("/produtos")
+@CrossOrigin(origins = "*", allowedHeaders = "*")
 public class ProdutoController {
+    
     @Autowired
     private ProdutoRepository produtoRepository;
-
+    
     @GetMapping
-    public List<Produto> listarTodos() {
-        return produtoRepository.findAll();
+    public ResponseEntity<List<Produto>> getAll(){
+        return ResponseEntity.ok(produtoRepository.findAll());
     }
-
+    
+    @GetMapping("/{id}")
+    public ResponseEntity<Produto> getById(@PathVariable Long id){
+        return produtoRepository.findById(id)
+            .map(resposta -> ResponseEntity.ok(resposta))
+            .orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
+    }
+    
+    @GetMapping("/nome/{nome}")
+    public ResponseEntity<List<Produto>> getByName(@PathVariable String nome){
+        return ResponseEntity.ok(produtoRepository
+        		.findAllByNomeContainingIgnoreCase(nome)); 
+        
+    }
+    
     @PostMapping
-    public Produto criar(@RequestBody Produto produto) {
-        return produtoRepository.save(produto);
+    public ResponseEntity<Produto> post(@Valid @RequestBody Produto produto){
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body(produtoRepository.save(produto));
     }
-
-    @PutMapping("/{id}")
-    public Produto atualizar(@PathVariable Long id, @RequestBody Produto produto) {
-        produto.setId(id);
-        return produtoRepository.save(produto);
+    
+    @PutMapping
+    public ResponseEntity<Produto> put(@Valid @RequestBody Produto produto){
+        return produtoRepository.findById(produto.getId())
+            .map(resposta -> ResponseEntity.status(HttpStatus.CREATED)
+            .body(produtoRepository.save(produto)))
+            .orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
     }
-
+    
+    @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/{id}")
-    public void deletar(@PathVariable Long id) {
-        produtoRepository.deleteById(id);
+    public void delete(@PathVariable Long id) {
+        Optional<Produto> produto = produtoRepository.findById(id);
+        
+        if(produto.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+        produtoRepository.deleteById(id);              
     }
+
 }
